+++ app/src/main/java/com/renobile/carrinho/domain/Product.kt
@@ -10,7 +10,7 @@ open class Product : RealmObject() {
     var cartId: Long = 0
     var listId: Long = 0
     var name: String = ""
-    var quantity: Int = 0
+    var quantity: Double = 0.0
     var price: Double = 0.0
 
 }
